+++ src/ui/camera.ts
@@ -1154,6 +1154,7 @@ export abstract class Camera extends Evented {
             offset: options.offset,
             zoom: options.zoom,
             center: options.center,
+            panInertia: (options as any)._panInertia,
         });
 
         this._rotating = this._rotating || (startBearing !== bearing);